compiliable with clang & split Shapes
