+++ CustomLibrary/CustomLibrary/SDL/Renderer.h
@@ -1,7 +1,8 @@
 #pragma once
 
-#include <CustomLibrary/Error.h>
-#include <CustomLibrary/Traits.h>
+#include "../Error.h"
+#include "../Traits.h"
+#include "Dim.h"
 #include <SDL.h>
 
 #include <cassert>
@@ -19,6 +20,13 @@ namespace ctl::sdl
 				throw ctl::Log(SDL_GetError());
 
 			SDL_SetRenderDrawColor(m_renderer, 0xFF, 0xFF, 0xFF, 0xFF);
+		}
+
+		basicRenderer(const basicRenderer&) = delete;
+		basicRenderer(basicRenderer&& r) noexcept
+			: m_renderer(r.m_renderer)
+		{
+			r.m_renderer = nullptr;
 		}
 
 		~basicRenderer()
@@ -48,7 +56,7 @@ namespace ctl::sdl
 	class ERendererDefault : public crtp<Impl, ERendererDefault>
 	{
 	public:
-		void logical_size(const Dim<int>& dim)
+		void logical_size(const ctl::sdl::Dim<int>& dim)
 		{
 			assert(this->_().get() != nullptr && "Renderer isn't loaded.");
 			SDL_RenderSetLogicalSize(this->_().get(), dim.w, dim.h);
